Orig rev 29
git-svn-id: https://time4u.svn.sourceforge.net/svnroot/time4u/overhaul/trunk@34 d1e209a9-466a-4101-bc37-f2a0db2545eb

+++ server/api/src/main/java/de/objectcode/time4u/server/api/data/ISynchronizableData.java
@@ -21,5 +21,5 @@ public interface ISynchronizableData extends Serializable
    * 
    * @return The revision number
    */
-  int getRevision();
+  long getRevision();
 }
